[Fix #3209] Remove faulty line length check from `Style/GuardClause` cop
This cop would attempt to predict how long the line would be after
turning an `if`- or `unless` statement into a guard clause.

Unfortunately, in many cases, this would reward poorly factored code
that produced longer guard clauses.

It would also cause a large amount of false negatives. For example:

```
if rand_one?(node)
  add_offense(node, :expression, format(MSG, node.source))
end
```

would, for reasons unknown, not generate an offense.