Test against up-to-date MRI's in Travis
We're are relying on rvm's fuzzy matching, but only for some rubies,
since the version of rvm provided by Travis is not up to date enough to
be able to fuzzy match all rubies.

Also, it doesn't fuzzy match very good. For example, it's choosing 2.2.0
over 2.2.5. One would expect it to choose the latest.

Given all of the above, I think it's better to manually keep fully
qualified versions.