[Fix #3578] Fix safe navigation counting in `Metrics/AbcSize` (#3581)
This cop wasn't correctly counting method calls that happen with the safe navigation operator (`x&.y`), it has been changed so that they are treated the same as other method calls (like `x.y`).

Adds a test which checks the score of a method with the safe navigation operator. The test would fail without the change in the cop.