[Fix #2731] Customise method-creating methods (#3883)
This addresses the issue raised in #2731 where methods (often from outside of Ruby itself) may be called within the context of a module that in turn define methods. The `Lint/UselessAccessModifier` wasn’t taking these into account.

Much like the existing approach taken with blocks/context-creating methods, there’s now a configuration option `MethodCreatingMethods` for the cop. This setting is an array of methods which, when called, are known to create other methods in the module’s current access context.