Add `default` key in `Style/PercentLiteralDelimiters` (#4038)
Oftentimes people want to set the default percent literal delimiter to [], but have to override every available option in order to get that behavior. This commit adds the ability to specify a default key that will specify the default for every possible delimiter, which you can then overwrite with others.